Start to change Contact.tags to a nested object list of ResourceTag
