+++ src/backend/main/py.main/caliopen_main/user/parameters/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import, print_function, unicode_literals
 
-from .user import NewUser, User, NewTag, Tag, NewRule
+from .user import NewUser, User,NewRule
 
 from .types import InternetAddressType, PhoneNumberType
 
@@ -10,13 +10,15 @@ from .contact import NewPostalAddress, PostalAddress
 from .contact import NewEmail, Email, NewIM, IM, NewPhone, Phone
 from .contact import NewSocialIdentity, SocialIdentity, NewPublicKey, PublicKey
 from .contact import NewContact, Contact, ShortContact
+from .tag import ResourceTag, NewUserTag, UserTag
 
 __all__ = [
     'InternetAddressType', 'PhoneNumberType',
-    'NewUser', 'User', 'NewTag', 'Tag', 'NewRule',
+    'NewUser', 'User', 'NewRule',
     'Recipient', 'NewOrganization', 'Organization',
     'NewPostalAddress', 'PostalAddress',
     'NewEmail', 'Email', 'NewIM', 'IM', 'NewPhone', 'Phone',
     'NewSocialIdentity', 'SocialIdentity', 'NewPublicKey', 'PublicKey',
-    'NewContact', 'Contact', 'ShortContact'
+    'NewContact', 'Contact', 'ShortContact',
+    'NewUserTag', 'UserTag', 'ResourceTag',
 ]
